back_end/requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SocketIO==5.3.7
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
paho-mqtt==2.1.0
psycopg2==2.9.9
python-can==4.4.2
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
typing_extensions==4.12.2
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0